Type switch for dimensions

+++ rt-one-week/Renderer.cpp
@@ -6,26 +6,26 @@
 
 Renderer::Renderer() : width(DEFAULT_WIDTH), height(DEFAULT_HEIGHT) {};
 
-Renderer::Renderer(const size_t width, const size_t height) : width(width), height(height) {}
-void Renderer::setWidth(const size_t width)
+Renderer::Renderer(const unsigned width, const unsigned height) : width(width), height(height) {}
+void Renderer::setWidth(const unsigned width)
 {
 	this->width = width;
 }
 
-void Renderer::setHeight(const size_t height)
+void Renderer::setHeight(const unsigned height)
 {
 	this->height = height;
 }
 
 Renderer::~Renderer() {}
 
-size_t Renderer::getHeight()
+unsigned Renderer::getHeight()
 {
 	return this->height;
 }
 
 
-size_t Renderer::getWidth()
+unsigned Renderer::getWidth()
 {
 	return this->width;
 }
