+++ rt-one-week/Renderer.h
@@ -7,16 +7,16 @@ constexpr auto DEFAULT_HEIGHT = 768;;
 class Renderer
 {
 private:
-	size_t width;
-	size_t height;
+	unsigned width;
+	unsigned height;
 public:
 	Renderer();
-	Renderer(const size_t, const size_t);
+	Renderer(const unsigned, const unsigned);
 	
-	void setWidth(const size_t);
-	void setHeight(const size_t);
-	size_t getWidth();
-	size_t getHeight();
+	void setWidth(const unsigned);
+	void setHeight(const unsigned);
+	unsigned getWidth();
+	unsigned getHeight();
 	void writeHeader(std::ostream& os);
 
 	~Renderer();
